refactor: Update OnchainKit and layout components
This commit refactors the code by updating the OnchainKit library to version 0.33.6 and making changes to the layout component. The OnchainProviders component is now imported directly in the layout component, removing the need for dynamic import. The metadata for the app has also been updated to reflect the new app name and description. Additionally, the WagmiProvider component has been updated to use the wagmiAdapter's wagmiConfig. The Providers component has been removed as it is no longer needed.